+++ src/tests/users.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UsersService } from './users.service';
+import { UsersService } from '../users/users.service';
 
 describe('UsersService', () => {
   let service: UsersService;
